resumax_backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
beautifulsoup4==4.13.3
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
defusedxml==0.7.1
Django==5.1.3
django-bootstrap-icons==0.9.0
django-environ==0.12.0
djangorestframework==3.15.2
dnspython==2.7.0
email_validator==2.2.0
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.5.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.1
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
pytest==8.3.4
pytest-django==4.10.0
requests==2.32.3
rsa==4.9
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
websockets==14.2
